+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 3.0.4 (TBD)
 
 * simple-features-java version 2.0.4
+* Projection Factory projection by name method
 
 ## [3.0.3](https://github.com/ngageoint/simple-features-proj-java/releases/tag/3.0.3) (07-13-2020)